% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenoptrExamples.R
\docType{package}
\name{phenoptrExamples-package}
\alias{phenoptrExamples}
\alias{phenoptrExamples-package}
\title{Example data for use with the \code{phenoptr} package}
\description{
\code{phenoptrExamples} contains data sets and example code which demonstrate
the use of the \code{phenoptr} package. \code{phenoptr} contains functions that make
it easier to read and analyze data tables
and images created by Akoya Biosciences' inForm software.
}
\details{
\code{phenoptrExamples} contains selected inForm output from nine images taken
from three samples of lung cancer tissue. To see the names of all of the
files included for these samples, use the command
\code{list.files(system.file("extdata", "samples", package = "phenoptrExamples"))}.

For details on the file contents, see the
\href{https://akoyabio.github.io/phenoptrExamples/}{package documentation}
online.

\code{phenoptr} and \code{phenoptrExamples} are part of the Akoya Biosciences
Phenoptics family of
Quantitative Pathology Research Solutions. For more information visit the
Phenoptics \href{http://www.perkinelmer.com/cancer-immunology/index.html}{home page}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/akoyabio/phenoptrExamples}
  \item Report bugs at \url{https://github.com/akoyabio/phenoptrExamples/issues}
}

}
\author{
\strong{Maintainer}: Kent Johnson \email{kjohnson@akoyabio.com}

Other contributors:
\itemize{
  \item Akoya Biosciences [copyright holder]
}

}
