% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenoptrExamples.R
\docType{package}
\name{phenoptrExamples-package}
\alias{phenoptrExamples}
\alias{phenoptrExamples-package}
\title{Example data for use with the \code{phenoptr} package}
\description{
\code{phenoptrExamples} contains data sets and example code which demonstrate
the use of the \code{phenoptr} package. \code{phenoptr} contains functions that make
it easier to read and analyze data tables
and images created by PerkinElmer's inForm software.
}
\details{
\code{phenoptr} and \code{phenoptrExamples} are part of the PerkinElmer Phenoptics family of
Quantitative Pathology Research Solutions. For more information visit the
Phenoptics \href{http://www.perkinelmer.com/cancer-immunology/index.html}{homepage}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/PerkinElmer/phenoptrExamples}
  \item Report bugs at \url{https://github.com/PerkinElmer/phenoptrExamples/issues}
}

}
\author{
\strong{Maintainer}: Kent Johnson \email{kent.johnson@perkinelmer.com}

Other contributors:
\itemize{
  \item PerkinElmer, Inc. [copyright holder]
}

}
